requirements.txt
fastapi
uvicorn
motor
pydantic
langchain
langchain-openai
python-dotenv